ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
Pillow>=10.0.0
joblib>=1.3.0
# Menggunakan scikit-image untuk image processing canggih tanpa dependensi berat TF/Torch
scikit-image>=0.21.0
requests>=2.31.0
